==> Codes/1_run_cv.rmd <==
# Setup

```{r echo = F, include = FALSE, cache = F}
library(knitr)
library(tidyverse)
opts_chunk$set(
  fig.align = "center",
  fig.retina = 5,
  warning = F,
  message = F,
  cache = T,
  echo = F
)

#* source functions
fs::dir_ls(here::here("Codes/functions")) %>%
  lapply(., source)
```

# Prepare dataset

```{r }
set.seed(2312390)
data_for_cv <-
  readRDS(here("Data/all_data.rds")) %>%
  #* generate data for spatial cross-validation
  mutate(split_data = list(
    spatial_clustering_cv(
      data,
      coords = c("Y", "X"),
      v = 10 # 10-fold
    )
  )) %>%
  #* find a set of variables that are not very highly
  #* correlated with each other
  mutate(vars_with_low_cor = list(
    find_vars_to_keep(data[, ..x_vars_to_keep])
  ))

saveRDS(data_for_cv, "Data/data_for_cv.rds")
```

# Spatial CV of RF

```{r }
set.seed(893423)

RF_sp_cv <-
  data_for_cv %>%
  #* create a table of hyper parameters to tune
  mutate(tune_pars = list(
    expand.grid(
      mtry = c(3, 5),
      min.node.size = c(5, 10, 15),
      sample.fraction = c(0.3, 0.4, 0.5)
    ) %>%
      expand_grid_df(
        .,
        tibble(
          vars_set = list(x_vars_to_keep, vars_with_low_cor),
          short = c(FALSE, TRUE)
        )
      ) %>%
      data.table()
  )) %>%
  #* spatial cross validation
  mutate(cv_results = list(
    mclapply(
      seq_len(nrow(tune_pars)),
      function(x) {
        print(x)

        x_vars <- c("nrate", tune_pars[x, vars_set][[1]])
        formula_yield <-
          paste0(
            "yield ~ ",
            paste0(x_vars, collapse = "+")
          )

        cv_data <-
          run_sptial_cv_rf(
            split_data = split_data,
            formula = formula_yield,
            mtry = tune_pars[x, mtry],
            min.node.size = tune_pars[x, min.node.size],
            sample.fraction = tune_pars[x, sample.fraction]
          ) %>%
          .[, `:=`(
            mtry = tune_pars[x, mtry],
            min.node.size = tune_pars[x, min.node.size],
            sample.fraction = tune_pars[x, sample.fraction],
            x_vars = tune_pars[x, vars_set],
            formula_yield = formula_yield,
            short = tune_pars[x, short]
          )]
        return(cv_data)
      },
      mc.cores = 9
    ) %>%
      rbindlist()
  )) %>%
  dplyr::select(id_field, cv_results) %>%
  unnest() %>%
  data.table()

saveRDS(RF_sp_cv, here("Results/rf_sp_cv.rds"))
```

# Spatial CV of GAM

```{r }
# split_data <- gam_sp_cv$split_data[[1]]
# vars_ls <- gam_sp_cv$vars_ls[[1]]
# x <- vars_ls[[1]]

gam_sp_cv <-
  data_for_cv %>%
  #* parameters to tune (set of variables)
  mutate(tune_pars = list(
    data.table(
      vars_set = list(x_vars_to_keep, vars_with_low_cor),
      short = c(FALSE, TRUE)
    )
    #* spatial cross validation
  )) %>%
  mutate(cv_results = list(
    mclapply(
      seq_len(nrow(tune_pars)),
      function(x) {
        print(id_field)

        #* vairables to use
        vars <- tune_pars[x, vars_set][[1]]

        #* gam formula
        gam_formula <-
          paste0(
            "yield ~ s(nrate, k = 2) + ",
            paste0("I(nrate*", vars, ")", collapse = " + "), " + ",
            paste0("s(", vars, ", k = 3)", collapse = " + ")
          )

        #* cross-validation
        cv_data <-
          run_sptial_cv_gam(
            split_data,
            gam_formula = gam_formula
          ) %>%
          .[, `:=`(
            x_vars = list(vars),
            short = tune_pars[x, short],
            gam_formula = list(gam_formula)
          )]

        return(cv_data)
      },
      mc.cores = 2
    ) %>%
      rbindlist()
  )) %>%
  dplyr::select(id_field, cv_results) %>%
  unnest() %>%
  data.table()

saveRDS(gam_sp_cv, here("Results/gam_sp_cv.rds"))
```
