==> Codes/2_train_model.rmd <==
# Setup

```{r echo = F, include = FALSE, cache = F}
library(knitr)
library(tidyverse)
opts_chunk$set(
  fig.align = "center",
  fig.retina = 5,
  warning = F,
  message = F,
  cache = T,
  echo = F
)

#* source functions
fs::dir_ls(here::here("Codes/functions")) %>%
  lapply(., source)
```

# Prepare dataset

```{r }
difm_geo <- readRDS(here("Data/all_data.rds"))
```

# Train RF

```{r }
rf_sp_cv <-
  readRDS("Results/rf_sp_cv.rds") %>%
  # .[short == TRUE, ] %>%
  #* pick the best hyper parametres and variable set
  .[, .SD[which.min(rmse_cv), ], by = id_field]

set.seed(782433)

rf_results <-
  difm_geo %>%
  left_join(., rf_sp_cv, by = "id_field") %>%
  mutate(all_x_vars = list(c("nrate", x_vars))) %>%
  mutate(rf_trained = list(
    ranger(
      formula = formula_yield,
      data = data,
      mtry = mtry,
      min.node.size = min.node.size,
      sample.fraction = sample.fraction,
      num.threads = 1,
      num.trees = 1000
    )
  )) %>%
  mutate(n_seq = list(
    data[, seq(min(nrate), max(nrate), by = 3)]
  )) %>%
  mutate(eonr_data = list(
    copy(data) %>%
      .[, nrate := NULL] %>%
      expand_grid_df(., data.table(nrate = n_seq)) %>%
      .[, y_hat :=
        predict(
          rf_trained,
          data = .
        )$predictions] %>%
      .[, pi_hat := Pw * y_hat - Pn * nrate] %>%
      .[, .SD[which.max(pi_hat), ], by = obs_id]
  )) %>%
  mutate(
    rmse_train =
      eonr_data[, (yield - y_hat)^2 %>% mean() %>% sqrt()]
  ) %>%
  mutate(
    r2_train =
      lm(yield ~ y_hat, data = eonr_data) %>%
        summary() %>%
        .$r.squared
  )

saveRDS(rf_results, "Results/rf_results.rds")

rf_results %>%
  dplyr::select(id_field, eonr_data, x_vars) %>%
  saveRDS("Results/rf_eonr_results.rds")

#* save rmse for a figure
rf_results %>%
  dplyr::select(id_field, rmse_cv, rmse_train) %>%
  saveRDS("Results/rf_rmse.rds")

# mutate(eonr_driver = list(
#   regression_forest(
#     X = eonr_data[, ..x_vars],
#     Y = eonr_data[, nrate],
#     num.threads = 1,
#     num.trees = 1000
#   )
# )) %>%
# mutate(var_i_eonr = list(
#   variable_importance(eonr_driver) %>%
#     data.table() %>%
#     .[, term := x_vars]
# )) %>%
# mutate(eonr_feols = list(
#   feols(nrate ~ .[x_vars], data = eonr_data) %>%
#     tidy() %>%
#     data.table() %>%
#     .[, .(term, estimate, statistic)]
# )) %>%
# mutate(sd_data = list(
#   eonr_data[, ..x_vars] %>%
#     .[
#       ,
#       lapply(
#         .SD,
#         function(x) {
#           quantile(x, prob = 0.75) - quantile(x, prob = 0.5)
#         }
#       )
#     ] %>%
#     melt() %>%
#     setnames("variable", "term")
# )) %>%
# mutate(eonr_impact = list(
#   sd_data[eonr_feols, on = "term"] %>%
#     .[, mi_impact := estimate * value]
# )) %>%
# mutate(var_i_eonr = list(
#   eonr_impact[var_i_eonr, on = "term"]
# )) %>%
# dplyr::select(-eonr_impact, -sd_data, -eonr_feols)
```


# Train GAM

```{r }
gam_sp_cv <-
  readRDS("Results/gam_sp_cv.rds") %>%
  .[short == TRUE, ] %>%
  #* pick the best hyper parametres and variable set
  .[, .SD[which.min(rmse_cv), ], by = id_field]

gam_results <-
  difm_geo %>%
  left_join(., gam_sp_cv, by = "id_field") %>%
  mutate(gam_trained = list(
    mgcv::gam(
      formula(gam_formula),
      data = data,
      method = "REML"
    )
  )) %>%
  mutate(n_seq = list(
    data[, seq(min(nrate), max(nrate), by = 3)]
  )) %>%
  mutate(eonr_data = list(
    copy(data) %>%
      .[, nrate := NULL] %>%
      expand_grid_df(., data.table(nrate = n_seq)) %>%
      .[, y_hat := predict(gam_trained, newdata = .)] %>%
      .[, pi_hat := Pw * y_hat - Pn * nrate] %>%
      .[, .SD[which.max(pi_hat), ], by = obs_id]
  )) %>%
  mutate(
    rmse_train =
      eonr_data[, (yield - y_hat)^2 %>% mean() %>% sqrt()]
  ) %>%
  mutate(
    r2_train =
      lm(yield ~ y_hat, data = eonr_data) %>%
        summary() %>%
        .$r.squared
  )

saveRDS(gam_results, "Results/gam_results.rds")

gam_results %>%
  dplyr::select(id_field, eonr_data, x_vars) %>%
  saveRDS("Results/gam_eonr_results.rds")

gam_results %>%
  dplyr::select(id_field, rmse_cv, rmse_train) %>%
  saveRDS("Results/gam_rmse.rds")
```

# Shapley values for yield prediction

```{r }
rf_results <- readRDS(here("Results/rf_results.rds"))

shap_yield_rf <-
  rf_results %>%
  dplyr::select(id_field, x_vars, data, rf_trained) %>%
  mutate(shap_values = list(
    get_shap_values(
      data = data,
      x_vars = c("nrate", x_vars),
      trained_model = rf_trained
    )
  )) %>%
  dplyr::select(-rf_trained, -data, -x_vars) %>%
  unnest() %>%
  data.table()

saveRDS(shap_yield_rf, here("Results/shap_yield_rf.rds"))
```

```{r }
gam_results <- readRDS(here("Results/gam_results.rds"))

shap_yield_gam <-
  gam_results %>%
  dplyr::select(id_field, x_vars, data, gam_trained) %>%
  .[1, ] %>%
  mutate(shap_values = list(
    get_shap_values(
      data = data,
      x_vars = c("nrate", x_vars),
      trained_model = gam_trained
    )
  )) %>%
  dplyr::select(-gam_trained, -data, -x_vars) %>%
  unnest() %>%
  data.table()

saveRDS(shap_yield_gam, here("Results/shap_yield_gam.rds"))
```
